+++ packages/experiments/src/demo/index.tsx
@@ -9,6 +9,7 @@ import { GettingStartedPage } from './GettingStartedPage';
 import { setBaseUrl } from 'office-ui-fabric-react/lib/Utilities';
 import { Fabric } from 'office-ui-fabric-react/lib/Fabric';
 import { initializeIcons } from '@uifabric/icons/lib/index';
+import { initializeFileTypeIcons } from '@uifabric/file-type-icons/src/index';
 
 import './index.scss';
 import './ColorStyles.scss';
@@ -16,7 +17,8 @@ import './ColorStyles.scss';
 setBaseUrl('./dist/');
 
 // Initialize all icons.
-initializeIcons('dist/');
+initializeIcons();
+initializeFileTypeIcons();
 
 let rootElement: HTMLElement | null;