+++ packages/office-ui-fabric-react/src/components/ComboBox/ComboBox.types.ts
@@ -117,6 +117,16 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox>
    * Add additional content below the callout list.
    */
   onRenderLowerContent?: IRenderFunction<IComboBoxProps>;
+
+  /**
+  * Custom width for dropdown (unless useComboBoxAsMenuWidth is undefined or false)
+  */
+  dropdownWidth?: number;
+
+  /**
+   * Whether to use the ComboBoxes width as the menu's width
+   */
+  useComboBoxAsMenuWidth?: boolean;
 }
 
 export interface IComboBoxStyles {
@@ -155,6 +165,12 @@ export interface IComboBoxStyles {
    * the comboBox is disabled.
    */
   rootHovered: IStyle;
+
+  /**
+   * Styles for when the ComboBox is active. These styles are applied for all comboBoxes except when
+   * the comboBox is disabled.
+   */
+  rootPressed: IStyle;
 
   /**
    * Styles for when the ComboBox is focused. These styles are applied for all comboBoxes except when