+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.styles.ts
@@ -4,20 +4,19 @@ import {
   IRawStyle,
   ITheme,
   concatStyleSets,
-  getFocusStyle
+  getFocusStyle,
+  HighContrastSelector
 } from '../../Styling';
 import { IContextualMenuStyles, IMenuItemStyles } from './ContextualMenu.types';
 
 import { memoizeFunction } from '../../Utilities';
-
-const MS_HIGHCONTRAST_ACTIVE = '@media screen and (-ms-high-contrast: active)';
 
 const ContextualMenuItemHeight = '32px';
 
 const getItemHighContrastStyles = memoizeFunction((): IRawStyle => {
   return {
     selectors: {
-      [MS_HIGHCONTRAST_ACTIVE]: {
+      [HighContrastSelector]: {
         backgroundColor: 'Highlight',
         borderColor: 'Highlight',
         color: 'HighlightText',
@@ -87,7 +86,6 @@ export const getMenuItemStyles = memoizeFunction((
     rootExpanded: {
       backgroundColor: ContextualMenuItemBackgroundSelectedColor,
       color: semanticColors.bodyTextChecked,
-      fontWeight: FontWeights.semibold,
       ...getItemHighContrastStyles()
     },
     linkContent: {
@@ -130,7 +128,7 @@ export const getMenuItemStyles = memoizeFunction((
     iconColor: {
       color: semanticColors.menuIcon,
       selectors: {
-        [MS_HIGHCONTRAST_ACTIVE]: {
+        [HighContrastSelector]: {
           color: 'HighlightText',
         }
       }
@@ -138,7 +136,7 @@ export const getMenuItemStyles = memoizeFunction((
     checkmarkIcon: {
       color: semanticColors.bodySubtext,
       selectors: {
-        [MS_HIGHCONTRAST_ACTIVE]: {
+        [HighContrastSelector]: {
           color: 'HighlightText',
         }
       }